# 01_openapi_swagger_erp_pharma_v2.yaml
openapi: 3.0.3
info:
  title: ERP Pharmaceutique SaaS V2 API
  version: 1.0.0
  description: API REST pour ERP Pharmaceutique SaaS multi-tenant basé sur React, NestJS et Supabase PostgreSQL.
servers:
  - url: https://api.pharmaerp.local/api/v1
    description: Production / exemple
  - url: http://localhost:3000/api/v1
    description: Local development
security:
  - bearerAuth: []
...
paths:
...
  /articles:
    get:
      tags: [Articles]
      summary: Liste des articles
      x-permission: articles.read
      parameters:
        - in: query
          name: search
          schema: { type: string }
        - in: query
          name: categoryId
          schema: { type: string, format: uuid }
        - in: query
          name: page
          schema: { type: integer, default: 1 }
        - in: query
          name: limit
          schema: { type: integer, default: 50 }
      responses:
        '200':
          description: Liste articles
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaginatedArticles'
    post:
      tags: [Articles]
      summary: Créer article
      x-permission: articles.create
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateArticleDto'
      responses:
        '201':
          description: Article créé
  /articles/{id}:
    get:
      tags: [Articles]
      summary: Détail article
      x-permission: articles.read
      parameters:
        - $ref: '#/components/parameters/IdPath'
      responses:
        '200':
          description: Article
    patch:
      tags: [Articles]
      summary: Modifier article
      x-permission: articles.update
      parameters:
        - $ref: '#/components/parameters/IdPath'
      responses:
        '200':
          description: Article modifié
...
  /stocks/articles/{articleId}:
    get:
      tags: [Stocks]
      summary: Stock par article et lot
      x-permission: stocks.read
      parameters:
        - name: articleId
          in: path
          required: true
          schema: { type: string, format: uuid }
      responses:
        '200':
          description: Stock article
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    IdPath:
      name: id
      in: path
      required: true
      schema:
        type: string
        format: uuid
  schemas:
...
    CreateArticleDto:
      type: object
      required: [articleCode, commercialName]
      properties:
        articleCode: { type: string }
        commercialName: { type: string }
        dci: { type: string }
        categoryId: { type: string, format: uuid }
        subCategoryId: { type: string, format: uuid }
        formId: { type: string, format: uuid }
        routeId: { type: string, format: uuid }
        dosage: { type: string }
        atcCode: { type: string }
        prescriptionRequired: { type: boolean }
        barcode: { type: string }
        defaultStockMin: { type: number }
    PaginatedArticles:
      type: object
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/Article'
        total: { type: integer }
    Article:
      type: object
      properties:
        articleId: { type: string, format: uuid }
        articleCode: { type: string }
        commercialName: { type: string }
        dci: { type: string }
        stockAvailable: { type: number }
        sellingPrice: { type: number }
...